+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.1.14](https://github.com/dwmkerr/crosswords-js/compare/v0.1.13...v0.1.14) (2023-03-13)
+
+### Bug Fixes
+
+- Backspace and Delete key handling ([#31](https://github.com/dwmkerr/crosswords-js/issues/31)) ([1ccd96e](https://github.com/dwmkerr/crosswords-js/commit/1ccd96e50d60cf74a8537d10baba4420321c77d7))
+
+## [0.1.13](https://github.com/dwmkerr/crosswords-js/compare/v0.1.12...v0.1.13) (2023-02-21)
+
+### Bug Fixes
+
+- correct eslint config and resolve broken tests ([0bc040f](https://github.com/dwmkerr/crosswords-js/commit/0bc040f9fe175dcd6a0a53c258f97885d2ebc19d))
+- correct package name ([20610e9](https://github.com/dwmkerr/crosswords-js/commit/20610e970459ebf012b904861ef7f3da4c0c41e4))
+
+## [0.1.12](https://github.com/dwmkerr/crosswords-js/compare/v0.1.11...v0.1.12) (2023-02-20)
+
+### Bug Fixes
+
+- Selected cell advances to "next" cell after keypress. ([5c6ca90](https://github.com/dwmkerr/crosswords-js/commit/5c6ca90e1177ccd48376a309c8217bb194f2ce6d))
+
 ### [0.1.11](https://github.com/dwmkerr/crosswords-js/compare/v0.1.10...v0.1.11) (2020-05-26)